# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "(Quiz) Server API - Swagger",
    "description": "Account Admin: admin - 12345 <br /> <br /> Account User: user - 12345",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "definitions": {
...
    "QuestionDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "What is NodeJS?"
        },
        "isMutiple": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "QuestionCreateDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "What is NodeJS?"
        },
        "isMutiple": {
          "type": "boolean",
          "example": true
        },
        "quiz": {
          "type": "object",
          "example": {
            "id": 1
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/question/get-all": {
      "get": {
        "tags": ["Question Controller"],
        "summary": "Get the list of question",
        "operationId": "getAllQuestion",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully - List of question",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/question/get-one/{id}": {
      "get": {
        "tags": ["Question Controller"],
        "summary": "Get question by id",
        "produces": ["application/xml", "application/json"],
        "operationId": "getQuestionById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of question to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully - Question Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID! ID Must be a number"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Question not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/question/create": {
      "post": {
        "tags": ["Question Controller"],
        "summary": "Create question",
        "operationId": "createQuestion",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "QuestionDTO",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/QuestionCreateDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/definitions/QuestionCreateDTO"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/definitions/QuestionCreateDTO"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/definitions/QuestionCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuestionCreateDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuestionCreateDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuestionCreateDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuestionCreateDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/question/update": {
      "put": {
        "tags": ["Question Controller"],
        "summary": "Update question",
        "operationId": "updateQuestion",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "QuestionDTO",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/QuestionDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/definitions/QuestionDTO"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/definitions/QuestionDTO"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/definitions/QuestionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Question not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/question/delete/{id}": {
      "delete": {
        "tags": ["Question Controller"],
        "summary": "Delete question",
        "produces": ["application/xml", "application/json"],
        "operationId": "deleteQuestionById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of quesstion to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/QuestionDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Question not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}